Restructuring to have only relevant files for demo
