+++ resolvers/queries.ts
@@ -1,41 +1,31 @@
-import { RouterContext } from "../deps.ts";
-import { queryPsqlResults } from "../psql-query.ts";
-const querySql = "select * from user_tbl;";
-const Query = {
-  getUsers: async (parent: any, args: any, context: any, info: any) => {
-    if (Object.keys(args).length === 0 && args.constructor === Object) {
-      return await queryPsqlResults(querySql);
+import { psqlClient } from "../db/psql-connect.ts";
+import { removeLastChar } from "../util/cleanup.ts";
+
+export const Query = {
+  getUsers: async (_parent: unknown, args: any, context: any, info: any) => {
+    let sql = "";
+
+    let order = "";
+    if ("order" in args) {
+      order = 'ORDER BY "' + args.order.colName + '" ' + args.order.orderBy +
+        " ";
+    }
+    if ("limit" in args) {
+      order = " LIMIT " + args.limit.limit;
+    }
+
+    if (Object.keys(args).indexOf("where") === -1) {
+      sql = "select * from user_tbl " + order + ";";
     } else {
       let where = " WHERE ";
-      Object.keys(args).map((key) =>
-        where += ' "' + key + '"' + "='" + args[key] + "' AND"
+      let inpt = args.where;
+      Object.keys(inpt).map((key) =>
+        where += ' "' + key + '"' + "='" + inpt[key] + "' AND"
       );
-
-      let res = where.split(" "); //split by space
-      res.pop(); //remove last element
-      let retstr = res.join(" ");
-      console.log("select * from user_tbl " + retstr + ";");
-      return await queryPsqlResults("select * from user_tbl " + retstr + ";");
-      //console.log(args['id']);
+      let retstr = removeLastChar(where);
+      sql = "select * from user_tbl " + retstr + " " + order + ";";
     }
-  },
-  Users: async (parent: any, args: any, context: any, info: any) => {
-    if (Object.keys(args).length === 0 && args.constructor === Object) {
-      return await queryPsqlResults(querySql);
-    } else {
-      let where = " WHERE ";
-      Object.keys(args).map((key) =>
-        where += ' "' + key + '"' + "='" + args[key] + "' AND"
-      );
-
-      let res = where.split(" "); //split by space
-      res.pop(); //remove last element
-      let retstr = res.join(" ");
-      console.log("select * from user_tbl " + retstr + ";");
-      return await queryPsqlResults("select * from user_tbl " + retstr + ";");
-      //console.log(args['id']);
-    }
+    const results = await psqlClient.queryObject(sql);
+    return results.rows;
   },
 };
-
-export { Query };
