+++ resolvers/resolvers.ts
@@ -1,7 +1,7 @@
-import { Query } from './queries.ts';
-import { Mutation } from './mutations.ts';
-// Resolvers
+import { Query } from "./queries.ts";
+import { Mutation } from "./mutations.ts";
+
 export const resolvers = {
-    Query,
-    Mutation
+  Query,
+  Mutation,
 };
